Separate out TestDevice and unix-specific devices
