+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["smoltcp-device"]
+members = ["smoltcp-device", "smoltcp-device-mock", "smoltcp-device-unix"]
 
 [package]
 name = "smoltcp"
@@ -42,18 +42,19 @@ rstest = "0.17"
 insta = "1.41.1"
 rand_chacha = "0.3.1"
 idna = { version = "=1.0.1" }
+smoltcp-device-mock = { path = "smoltcp-device-mock" }
+
+[target.'cfg(unix)'.dev-dependencies]
+smoltcp-device-unix = { path = "smoltcp-device-unix" }
 
 [features]
 std = ["managed/std", "alloc", "smoltcp-device/std"]
 alloc = ["managed/alloc", "defmt?/alloc"]
 verbose = []
 defmt = ["dep:defmt", "heapless/defmt", "smoltcp-device/defmt"]
-"medium-ethernet" = ["socket", "smoltcp-device/requires-medium-ethernet"]
-"medium-ip" = ["socket", "smoltcp-device/requires-medium-ip"]
-"medium-ieee802154" = ["socket", "proto-sixlowpan", "smoltcp-device/requires-medium-ieee802154"]
-
-"phy-raw_socket" = ["std", "libc", "smoltcp-device/provides-medium-ethernet", "smoltcp-device/provides-medium-ieee802154", "smoltcp-device/provides-medium-ip", "medium-ethernet", "medium-ieee802154", "medium-ip"]
-"phy-tuntap_interface" = ["std", "libc", "medium-ethernet"]
+"medium-ethernet" = ["socket", "smoltcp-device/requires-medium-ethernet", "smoltcp-device-mock/medium-ethernet", "smoltcp-device-unix/medium-ethernet"]
+"medium-ip" = ["socket", "smoltcp-device/requires-medium-ip", "smoltcp-device-mock/medium-ip", "smoltcp-device-unix/medium-ip"]
+"medium-ieee802154" = ["socket", "proto-sixlowpan", "smoltcp-device/requires-medium-ieee802154", "smoltcp-device-mock/medium-ieee802154", "smoltcp-device-unix/medium-ieee802154"]
 
 "proto-ipv4" = []
 "proto-ipv4-fragmentation" = ["proto-ipv4", "_proto-fragmentation"]
@@ -104,7 +105,6 @@ defmt = ["dep:defmt", "heapless/defmt", "smoltcp-device/defmt"]
 default = [
   "std", "log", # needed for `cargo test --no-default-features --features default` :/
   "medium-ethernet", "medium-ip", "medium-ieee802154",
-  "phy-raw_socket", "phy-tuntap_interface",
   "proto-ipv4",  "proto-dhcpv4", "proto-ipv6", "proto-dns",
   "proto-ipv4-fragmentation", "proto-sixlowpan-fragmentation",
   "socket-raw", "socket-icmp", "socket-udp", "socket-tcp", "socket-dhcpv4", "socket-dns", "socket-mdns",
@@ -117,6 +117,7 @@ default = [
 
 "_proto-fragmentation" = []
 
+# Prevents running netsim (slow) with the other tests.
 "_netsim" = []
 
 # BEGIN AUTOGENERATED CONFIG FEATURES
@@ -289,59 +290,59 @@ required-features = ["std"]
 
 [[example]]
 name = "tcpdump"
-required-features = ["std", "phy-raw_socket", "proto-ipv4"]
+required-features = ["std", "log", "medium-ethernet", "proto-ipv4"]
 
 [[example]]
 name = "httpclient"
-required-features = ["std", "medium-ethernet", "medium-ip", "phy-tuntap_interface", "proto-ipv4", "proto-ipv6", "socket-tcp"]
+required-features = ["std", "log", "medium-ethernet", "medium-ip", "proto-ipv4", "proto-ipv6", "socket-tcp"]
 
 [[example]]
 name = "ping"
-required-features = ["std", "medium-ethernet", "medium-ip", "phy-tuntap_interface", "proto-ipv4", "proto-ipv6", "socket-icmp"]
+required-features = ["std", "log", "medium-ethernet", "medium-ip", "proto-ipv4", "proto-ipv6", "socket-icmp"]
 
 [[example]]
 name = "server"
-required-features = ["std", "medium-ethernet", "medium-ip", "phy-tuntap_interface", "proto-ipv4", "socket-tcp", "socket-udp"]
+required-features = ["std", "log", "medium-ethernet", "medium-ip", "proto-ipv4", "socket-tcp", "socket-udp"]
 
 [[example]]
 name = "client"
-required-features = ["std", "medium-ethernet", "medium-ip", "phy-tuntap_interface", "proto-ipv4", "socket-tcp", "socket-udp"]
+required-features = ["std", "log", "medium-ethernet", "medium-ip", "proto-ipv4", "socket-tcp", "socket-udp"]
 
 [[example]]
 name = "loopback"
-required-features = ["log", "medium-ethernet", "proto-ipv4", "socket-tcp"]
+required-features = ["log", "medium-ethernet", "medium-ip", "medium-ieee802154", "proto-ipv4", "socket-tcp"]
 
 [[example]]
 name = "loopback_benchmark"
-required-features = ["std", "log", "medium-ethernet", "proto-ipv4", "socket-tcp"]
+required-features = ["std", "log", "medium-ethernet", "medium-ip", "medium-ieee802154", "proto-ipv4", "socket-tcp"]
 
 [[example]]
 name = "multicast"
-required-features = ["std", "medium-ethernet", "medium-ip", "phy-tuntap_interface", "proto-ipv4", "multicast", "socket-udp"]
+required-features = ["std", "log", "medium-ethernet", "medium-ip", "proto-ipv4", "multicast", "socket-udp"]
 
 [[example]]
 name = "multicast6"
-required-features = ["std", "medium-ethernet", "medium-ip", "phy-tuntap_interface", "proto-ipv6", "socket-udp"]
+required-features = ["std", "log", "medium-ethernet", "medium-ip", "proto-ipv6", "socket-udp"]
 
 [[example]]
 name = "benchmark"
-required-features = ["std", "medium-ethernet", "medium-ip", "phy-tuntap_interface", "proto-ipv4", "socket-raw", "socket-udp"]
+required-features = ["std", "log", "medium-ethernet", "medium-ip", "proto-ipv4", "socket-raw", "socket-udp"]
 
 [[example]]
 name = "dhcp_client"
-required-features = ["std", "medium-ethernet", "medium-ip", "phy-tuntap_interface", "proto-ipv4", "proto-dhcpv4", "socket-raw"]
+required-features = ["std", "log", "medium-ethernet", "medium-ip", "proto-ipv4", "proto-dhcpv4", "socket-raw"]
 
 [[example]]
 name = "sixlowpan"
-required-features = ["std", "medium-ieee802154", "phy-raw_socket", "proto-sixlowpan", "proto-sixlowpan-fragmentation", "socket-udp"]
+required-features = ["std", "log", "medium-ethernet", "medium-ieee802154", "medium-ip", "proto-sixlowpan", "proto-sixlowpan-fragmentation", "socket-udp"]
 
 [[example]]
 name = "sixlowpan_benchmark"
-required-features = ["std", "medium-ieee802154", "phy-raw_socket", "proto-sixlowpan", "proto-sixlowpan-fragmentation", "socket-udp"]
+required-features = ["std", "log", "medium-ethernet", "medium-ieee802154", "medium-ip", "proto-sixlowpan", "proto-sixlowpan-fragmentation", "socket-udp"]
 
 [[example]]
 name = "dns"
-required-features = ["std", "medium-ethernet", "medium-ip", "phy-tuntap_interface", "proto-ipv4", "socket-dns"]
+required-features = ["std", "log", "medium-ethernet", "medium-ip", "proto-ipv4", "socket-dns"]
 
 [profile.release]
 debug = 2
