+++ examples/server.rs
@@ -5,10 +5,11 @@ use std::fmt::Write;
 use std::os::unix::io::AsRawFd;
 
 use smoltcp::iface::{Config, Interface, SocketSet};
-use smoltcp::phy::{Device, Medium, wait as phy_wait};
+use smoltcp::phy::{Device, Medium};
 use smoltcp::socket::{tcp, udp};
 use smoltcp::time::{Duration, Instant};
 use smoltcp::wire::{EthernetAddress, IpAddress, IpCidr, Ipv4Address, Ipv6Address};
+use smoltcp_device_unix::wait as phy_wait;
 
 fn main() {
     utils::setup_logging("");
