+++ examples/sixlowpan_benchmark.rs
@@ -18,7 +18,7 @@
 //!
 //!
 //! # Setup
-//!    
+//!
 //!     modprobe mac802154_hwsim
 //!
 //!     ip link set wpan0 down
@@ -47,9 +47,10 @@ use std::os::unix::io::AsRawFd;
 use std::str;
 
 use smoltcp::iface::{Config, Interface, SocketSet};
-use smoltcp::phy::{Device, Medium, RawSocket, wait as phy_wait};
+use smoltcp::phy::{Device, Medium};
 use smoltcp::socket::tcp;
 use smoltcp::wire::{EthernetAddress, Ieee802154Address, Ieee802154Pan, IpAddress, IpCidr};
+use smoltcp_device_unix::{RawSocket, wait as phy_wait};
 
 //For benchmark
 use smoltcp::time::{Duration, Instant};
