+++ smoltcp-device-unix/src/macros.rs
@@ -1,0 +1,34 @@
+#![allow(unused_macros)]
+
+#[cfg(not(test))]
+#[cfg(feature = "log")]
+macro_rules! net_log {
+    (trace, $($arg:expr),*) => { log::trace!($($arg),*) };
+    (debug, $($arg:expr),*) => { log::debug!($($arg),*) };
+}
+
+#[cfg(test)]
+#[cfg(feature = "log")]
+macro_rules! net_log {
+    (trace, $($arg:expr),*) => { println!($($arg),*) };
+    (debug, $($arg:expr),*) => { println!($($arg),*) };
+}
+
+#[cfg(feature = "defmt")]
+macro_rules! net_log {
+    (trace, $($arg:expr),*) => { defmt::trace!($($arg),*) };
+    (debug, $($arg:expr),*) => { defmt::debug!($($arg),*) };
+}
+
+#[cfg(not(any(feature = "log", feature = "defmt")))]
+macro_rules! net_log {
+    ($level:ident, $($arg:expr),*) => {{ $( let _ = $arg; )* }}
+}
+
+macro_rules! net_trace {
+    ($($arg:expr),*) => (net_log!(trace, $($arg),*));
+}
+
+macro_rules! net_debug {
+    ($($arg:expr),*) => (net_log!(debug, $($arg),*));
+}
