+++ src/tests.rs
@@ -1,7 +1,7 @@
-use std::collections::VecDeque;
+pub(crate) use smoltcp_device_mock::TestingDevice;
 
 use crate::iface::*;
-use crate::phy::{self, Device, DeviceCapabilities, Medium};
+use crate::phy::Medium;
 use crate::time::Instant;
 use crate::wire::*;
 
@@ -51,97 +51,4 @@ pub(crate) fn setup<'a>(medium: Medium) -> (Interface, SocketSet<'a>, TestingDev
     }
 
     (iface, SocketSet::new(vec![]), device)
-}
-
-/// A testing device.
-#[derive(Debug)]
-pub struct TestingDevice {
-    pub(crate) tx_queue: VecDeque<Vec<u8>>,
-    pub(crate) rx_queue: VecDeque<Vec<u8>>,
-    max_transmission_unit: usize,
-    medium: Medium,
-}
-
-#[allow(clippy::new_without_default)]
-impl TestingDevice {
-    /// Creates a testing device.
-    ///
-    /// Every packet transmitted through this device will be received through it
-    /// in FIFO order.
-    pub fn new(medium: Medium) -> Self {
-        TestingDevice {
-            tx_queue: VecDeque::new(),
-            rx_queue: VecDeque::new(),
-            max_transmission_unit: match medium {
-                #[cfg(feature = "medium-ethernet")]
-                Medium::Ethernet => 1514,
-                #[cfg(feature = "medium-ip")]
-                Medium::Ip => 1500,
-                #[cfg(feature = "medium-ieee802154")]
-                Medium::Ieee802154 => 1500,
-            },
-            medium,
-        }
-    }
-}
-
-impl Device for TestingDevice {
-    type RxToken<'a> = RxToken;
-    type TxToken<'a> = TxToken<'a>;
-
-    fn capabilities(&self) -> DeviceCapabilities {
-        let mut capabilities = DeviceCapabilities::new(self.medium);
-
-        capabilities.max_transmission_unit = self.max_transmission_unit;
-
-        capabilities
-    }
-
-    fn receive(&mut self, _timestamp: Instant) -> Option<(Self::RxToken<'_>, Self::TxToken<'_>)> {
-        self.rx_queue.pop_front().map(move |buffer| {
-            let rx = RxToken { buffer };
-            let tx = TxToken {
-                queue: &mut self.tx_queue,
-            };
-            (rx, tx)
-        })
-    }
-
-    fn transmit(&mut self, _timestamp: Instant) -> Option<Self::TxToken<'_>> {
-        Some(TxToken {
-            queue: &mut self.tx_queue,
-        })
-    }
-}
-
-#[doc(hidden)]
-pub struct RxToken {
-    buffer: Vec<u8>,
-}
-
-impl phy::RxToken for RxToken {
-    fn consume<R, F>(self, f: F) -> R
-    where
-        F: FnOnce(&[u8]) -> R,
-    {
-        f(&self.buffer)
-    }
-}
-
-#[doc(hidden)]
-#[derive(Debug)]
-pub struct TxToken<'a> {
-    queue: &'a mut VecDeque<Vec<u8>>,
-}
-
-impl<'a> phy::TxToken for TxToken<'a> {
-    fn consume<R, F>(self, len: usize, f: F) -> R
-    where
-        F: FnOnce(&mut [u8]) -> R,
-    {
-        let mut buffer = vec![0; len];
-        let result = f(&mut buffer);
-        self.queue.push_back(buffer);
-        result
-    }
 }
